refactor(admin): move legacy KMS handlers to kms_keys
